Rename e2e to mock
