+++ bin/citrea/tests/mock/evm/ethers_js/package.json
