+++ bin/citrea/tests/mock/evm/uniswap/contracts/UsdCoin.sol
