+++ bin/citrea/tests/mock/evm/uniswap/scripts/01_deploy.js
