+++ bin/citrea/tests/mock/pruning.rs
