+++ bin/citrea/tests/mock_e2e.rs
@@ -1,7 +1,4 @@
 // disable bank module tests due to needing a big rewrite to make it work
 // mod bank;
 mod common;
-mod e2e;
-mod evm;
-mod mempool;
-mod soft_confirmation_rule_enforcer;
+mod mock;
